requirements.txt
asgiref==3.7.2
awscli==1.19.40
backports.zoneinfo==0.2.1
botocore==1.20.40
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
distlib==0.3.7
Django==4.2.8
djangorestframework==3.14.0
docutils==0.15.2
faiss-cpu==1.7.2
filelock==3.12.4
graphviz==0.8.4
idna==2.6
imageio==2.9.0
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.0
mxnet==1.7.0.post2
networkx==2.5
numpy==1.20.3
opencv-python==4.5.1.48
Pillow==8.1.2
platformdirs==3.11.0
pyasn1==0.4.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.16.0
pytz==2023.3.post1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.18.4
rsa==4.7.2
s3transfer==0.3.6
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
sqlparse==0.4.4
threadpoolctl==2.1.0
tifffile==2021.3.17
typing-extensions==4.8.0
tzdata==2023.3
urllib3==1.22
virtualenv==20.24.5
